\name{plot_tsne}
\alias{plot_tsne}
\title{Plot a t-SNE plot}
\description{This function plots a t-SNE from clustered migration data object.
}

\usage{
plot_tsne(MigrObj, uniq=NULL)
}

\arguments{
\item{MigrObj}{`MigrDat` object}
\item{uniq}{Identifier for clustering}
\item{color.by}{Variable to color dots}
\item{spots.slot}{Spots slot data}
\item{point.size}{Point size of dots}
}









