\name{plot_umap}
\alias{plot_umap}
\title{Plot a UMAP plot}
\description{This function plots a UMAP from clustered migration data object.
}

\usage{
plot_umap(MigrObj, uniq=NULL)
}

\arguments{
\item{MigrObj}{`MigrDat` object}
\item{uniq}{Identifier for clustering}
\item{color.by}{Variable to color dots}
\item{spots.slot}{Spots slot data}
\item{point.size}{Point size of dots}
}









