\name{plot_violin}
\alias{plot_violin}
\title{Plot a violin plot}
\description{
This function plots a violin plot of specific variable over clusters.
}

\usage{
plot_violin(MigrObj, feature = NULL, data.slot = "raw", type = "S", clusterType = "ILoRegclusters")
}

\arguments{
\item{MigrObj}{`MigrDat` object}
\item{feature}{Feature to be plotted}
\item{data.slot}{Data slot}
\item{type}{Type}
\item{clusterType}{Cluster type}
}









