\name{plot_rag}
\alias{plot_rag}
\title{Plot a Rag plot}
\description{This function plots a Rag carpet (Rag) plot from clustered migration data object.
}

\usage{
plot_rag(MigrObj, clusterType = "ILoRegclusters", feature = NULL)
}

\arguments{
\item{MigrObj}{`MigrDat` object}
\item{feature}{Numeric feature from the tracks table to be combined with TRACK_ID to create a new identifier for sorting the data}
\item{clusterType}{Type of clustering to use}
}









