\name{plot_rag}
\alias{plot_rag}
\title{Plot a Rag plot}
\description{This function plots a Rag carpet (Rag) plot from clustered migration data object.
}

\usage{
plot_rag(MigrObj, clusterType = "ILoRegclusters")
}

\arguments{
\item{MigrObj}{`MigrDat` object}
\item{clusterType}{Cluster type}
}









